requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
aiosqlite==0.19.0
python-dotenv==1.0.0
